requirements.txt
asgiref==3.3.4
bson==0.5.8
dataclasses==0.6
Django==2.2.22
django-cors-headers==3.7.0
djangorestframework==3.12.4
djongo==1.3.1
pymongo==3.11.4
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.2.4
